pipeline-manager: db: split off service implementation into own file
Split off the implementation of the service functions
for the database (new, list, get_by_id, get_by_name,
update, delete) into their own file. This helps with
readability by reducing the number of lines in the
mod file.
